Improvements to {Read,Write}Buffer classes and NamedPipe
 * Avoid std::stringstream because it's inefficient/bloated.

 * Explicitly handle decode errors, by throwing ReadBuffer::DecodeError.
   For the time being, this exception can only be thrown in the agent, but
   I anticipate it being thrown in the libwinpty code eventually.

 * Catch encoding errors a bit better.

 * Be more precise about the integer types used for length.  e.g.
   Use 64 bits on the wire, but otherwise use size_t/uint64_t where
   appropriate.

 * Replace memcpy with std::copy to make MSVC's /SDL security checks happy.
   It's a somewhat silly/arbitrary change.  I don't think it will affect
   performance, but I haven't benchmarked it.

 * This commit changes the RPC protocol that libwinpty uses to communicate
   with winpty-agent.

+++ src/agent/Terminal.cc
@@ -313,7 +313,7 @@ void Terminal::sendLine(int64_t line, const CHAR_INFO *lineData, int width)
     moveTerminalToLine(line);
 
     m_termLine.clear();
-    int trimmedLineLength = 0;
+    size_t trimmedLineLength = 0;
     bool alreadyErasedLine = false;
 
     int cellCount = 1;
